change type from REAL to INTEGER for rayleigh_type

+++ src/configure_model/mo_nonhydrostatic_config.f90
@@ -91,7 +91,7 @@ MODULE mo_nonhydrostatic_config
 
 
 
-    REAL(wp):: rayleigh_type    ! type of Rayleigh damping (1: CLASSIC, 2: Klemp (2008))
+    INTEGER :: rayleigh_type    ! type of Rayleigh damping (1: CLASSIC, 2: Klemp (2008))
     REAL(wp):: damp_height(max_dom)    ! height at which w-damping and sponge layer start
     REAL(wp):: rayleigh_coeff(max_dom) ! Rayleigh damping coefficient in w-equation
     REAL(wp):: vwind_offctr     ! Off-centering in vertical wind solver
